src/job_weather/gcp_distrb_crawler_weather/requirements.txt
retry-requests
cryptography
python-dotenv
pandas
requests
bs4
sqlalchemy
pymysql
pyarrow
google-cloud-storage
apache-airflow-providers-google
pendulum
graphviz